backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
chess==1.11.2
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.0
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
razorpay==2.0.0
requests==2.32.5
sgmllib3k==1.0.0
SQLAlchemy==2.0.46
sqlmodel==0.0.31
starlette==0.49.3
treys==0.1.8
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
